+++ tests/test_e2e.py
@@ -1,17 +1,16 @@
-import neptune.new as neptune
+import neptune
 import torch
-from neptune.new.integrations.sacred import NeptuneObserver
+
+from neptune_sacred import NeptuneObserver
 
 
 def test_e2e(experiment, model):
-    if torch.device("cuda:0"):
-        torch.cuda.empty_cache()
 
-    # Step 1: Initialize Neptune and create new Neptune Run
-    neptune_run = neptune.init_run()
+    # Step 1: Initialize Neptune and create new Neptune Handler
+    neptune_handler = neptune.init_run()["sacred"]
 
     # Step 2: Add NeptuneObserver() to your sacred experiment's observers
-    experiment.observers.append(NeptuneObserver(run=neptune_run))
+    experiment.observers.append(NeptuneObserver(run=neptune_handler))
 
     # Step 3: Run you experiment and explore metadata in Neptune UI
     experiment.run()
@@ -31,17 +30,18 @@ def test_e2e(experiment, model):
     experiment.add_artifact(name=model_fname + "_arch", filename=f"./{model_fname}_arch.txt")
     experiment.add_artifact(name=model_fname, filename=f"./{model_fname}.pth")
 
-    neptune_run.sync()
+    run = neptune_handler.get_root_object()
+    run.sync()
 
-    assert neptune_run["experiment/config/data_dir"].fetch() == "data/CIFAR10"
+    assert run["sacred/experiment/config/data_dir"].fetch() == "data/CIFAR10"
 
-    assert neptune_run.exists("experiment/io_files/artifacts/model")
-    assert neptune_run.exists("experiment/io_files/artifacts/model_arch")
+    assert run.exists("sacred/experiment/io_files/artifacts/model")
+    assert run.exists("sacred/experiment/io_files/artifacts/model_arch")
 
-    assert 0 < neptune_run["experiment/metrics/results/final_acc"].fetch() < 1
+    assert 0 < run["sacred/experiment/metrics/results/final_acc"].fetch() < 1
 
-    assert neptune_run.exists("experiment/sacred_config/experiment_info")
-    assert neptune_run.exists("experiment/sacred_config/meta_info")
+    assert run.exists("sacred/experiment/sacred_config/experiment_info")
+    assert run.exists("sacred/experiment/sacred_config/meta_info")
 
     # Stop run
-    neptune_run.stop()
+    run.stop()